Add VSCode task to start DDS tunnel
Just for convenience, I've made the DDS tunnel script executable and
added a VSCode task to start it.

It doesn't have to be executable for the VSCode task to be able to run
it, but that's just convenient since then you don't have to type
`python3 <file>`. Of course, this won't work on windows though.